refactor: 깃허브 스트릭, 비즈니스 로직 분리
